// src/app/models/survey-storage.model.ts
export class SurveyStorageModel {
  surveyMenu: any;
  canvasImage: string;
  formData: any;
  saved: boolean;

  surveyid: number;
  surveytype: string;
  city: string;
  state: string;
  latitude: number;
  longitude: number;
  menuitems : any;
  formdata: any;
  currentprogress : number;
  selectedmainmenuindex : number;
  selectedsubmenuindex : number;
  selectedshotindex : number;
  shotcompletecount : number;
  previousmainmenuindex : number;
  previoussubmenuindex : number;
  previousshotindex : number;
  remainingfilestoupload : number;
}