// src/app/models/constants.ts
export class ImageUploadModel {
  key: string;
  imageData: string;
  imagename: string;
}